Verifica si una persona es mayor de edad

Esta función determina si una persona es mayor de 18 años a partir de su fecha de nacimiento.

Usage

es_mayor_de_edad(fecha_nacimiento)

Arguments

fecha_nacimiento

Una cadena de texto con la fecha de nacimiento en formato "AAAA-MM-DD".

Value

`TRUE` si la persona tiene 18 años o más, `FALSE` en caso contrario.