Validar los datos de un contrato

Verifica que los datos de un contrato sean válidos, el formato correcto de fechas, la coherencia de fechas de entrada y salida, y la validez de otros campos como número de personas, número de habitaciones, tipo de pago e internet.

Usage

validar_contrato(comunicacion, contrato)

Arguments

comunicacion

Un `data.frame` que contiene las claves primarias de comunicación.

contrato

Un `data.frame` con la información de los contratos a validar.

Value

Devuelve `TRUE` si no se encuentran errores. En caso de valores inválidos, muestra advertencias.

Details

La validación incluye: - `referencia`: Debe estar presente. - `fechaContrato`: Debe cumplir con el formato de fecha `AAAA-MM-DD`. - `fechaEntrada` y `fechaSalida`: Deben cumplir con el formato `AAAA-MM-DDThh:mm:ss` y la fecha de salida debe ser posterior a la de entrada. - `numPersonas`: Debe ser un número entero mayor que 0. - `numHabitaciones`: Si está presente, debe ser un número entero mayor que 0. - `internet`: Si está presente, solo puede contener los valores `"true"` o `"false"`. - `tipoPago`: Debe ser un valor válido según la tabla `tipo_pago`. - `fechaPago`: Si está presente, debe cumplir con el formato `AAAA-MM-DD`. - `caducidadTarjeta`: Si está presente, debe cumplir con el formato `MM/AAAA`.

Si alguna validación falla, se generarán advertencias (`warning`) indicando el problema y la fila afectada.