Validar datos de personas en la comunicación

Valida los datos de las personas asociadas a una comunicación, asegurando que los campos obligatorios estén presentes y que la información proporcionada sea consistente. Genera advertencias en caso de datos faltantes o incorrectos.

Usage

validar_persona(comunicacion, persona)

Arguments

comunicacion

Data frame con los datos de la comunicación.

persona

Data frame con los datos de las personas asociadas a la comunicación.

Value

`TRUE` si todas las validaciones se completan sin problemas, de lo contrario, emite advertencias (`warning`).

Details

- Verifica que el campo `rol` solo tome los valores "TI" (Titular) o "VI" (Viajero). - Asegura que al menos una persona tenga el rol de Titular (TI). - Si hay menores de edad, al menos un adulto debe tener informado su parentesco con ellos. - Valida la presencia de campos obligatorios como `nombre`, `apellido1` y `fechaNacimiento`. - Si la persona es mayor de edad, se verifica la existencia de `tipoDocumento` y `numeroDocumento`. - Comprueba que `tipoDocumento` tenga un valor válido de `tipo_documento$codigo`. - Si el `tipoDocumento` es "NIF", el campo `apellido2` es obligatorio. - Si el `tipoDocumento` es "NIF" o "NIE", el campo `soporteDocumento` debe existir y tener 9 caracteres alfanuméricos. - Valida que `sexo`, `nacionalidad` y `parentesco` contengan valores válidos según las tablas de códigos. - Se debe proporcionar al menos un dato de contacto: `telefono`, `telefono2` o `correo`. - Se realizan validaciones sobre dirección, teléfono, email y documento de identidad.